pyproject.toml
[project]
name = "modal-rbe-backend"
version = "0.1.0"
description = "Bazel Remote Execution / Remote Cache backend backed by Modal"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "modal>=0.64",
    "grpcio>=1.60",
    "grpcio-status>=1.60",
    "protobuf>=4.25",
]

[project.optional-dependencies]
dev = [
    "grpcio-tools>=1.60",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modal_rbe"]
